% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\docType{data}
\name{.IPEA_BASE_URL}
\alias{.IPEA_BASE_URL}
\title{API Configuration and Data Fetching Functions}
\format{
An object of class \code{character} of length 1.
}
\usage{
.IPEA_BASE_URL
}
\description{
Functions to handle API interactions with IPEA's database for price index data.
}
\keyword{datasets}
