% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{get_cache_file_path}
\alias{get_cache_file_path}
\title{Get Cache File Path for Index}
\usage{
get_cache_file_path(index)
}
\arguments{
\item{index}{Character string indicating the price index.}
}
\value{
Character string with the full path to the cache file.
}
\description{
Returns the full path to the cache file for a specific price index.
}
