% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipca.R
\name{ipca}
\alias{ipca}
\title{Deflate Brazilian Reais in Nominal Values Using IPCA}
\usage{
ipca(data, values, actual_date, ref_date, outname = "def_values")
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble} containing nominal Reais and their original dates;}

\item{values}{A vector containing nominal Brazilian Reais to deflate;}

\item{actual_date}{A vector of class \code{Date} containing the nominal values' original dates;}

\item{ref_date}{An object of class \code{Date} indicating the reference date to deflate nominal values
(e.g., 2018-01-01 for January 2018).}

\item{outname}{A \code{character} specifying the name of the output variable with current Reais.}
}
\value{
A \code{numeric} vector of current Brazilian Reais.
}
\description{
\code{ipca} uses data from the Brazilian Institute of Geography and Statistics (IBGE) to deflate
Brazilian Reais nominal values using the IPCA to current values. As of now, the \code{ipca} deflate
values by month
}
\details{
The data used to deflate nominal values comes from the IBGE's Sidra, convering the period
from December 1979 up to now.
}
\examples{
\dontrun{
# Use IPCA to deflate a vector of nominal Brazilian Reais
reais <- rep(100, 5)
actual_dates <- seq.Date(from = as.Date("2001-01-01"), to = as.Date("2001-05-01"), by = "month")

ipca(reais, actual_dates, as.Date("2018-01-01"))
}

}
\references{
For more information on the IBGE's data, please check (in Portuguese):
\url{https://sidra.ibge.gov.br/pesquisa/snipc}.
}
