% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{save_to_cache}
\alias{save_to_cache}
\title{Save Data to Cache}
\usage{
save_to_cache(index, data, verbose = TRUE)
}
\arguments{
\item{index}{Character string indicating the price index.}

\item{data}{Data frame with price index data to cache.}

\item{verbose}{Logical indicating whether to show cache saving messages.}
}
\value{
Logical indicating success of the caching operation.
}
\description{
Saves price index data to the local cache file.
}
